Refactor code using useLoadCategories hook
